+++ pg.el
@@ -3,7 +3,7 @@
 ;; Copyright: (C) 1999-2002, 2022  Eric Marsden
 
 ;; Author: Eric Marsden <eric.marsden@risk-engineering.org>
-;; Version: 0.18
+;; Version: 0.19
 ;; Keywords: data comm database postgresql
 ;; URL: https://github.com/emarsden/pg-el
 ;; Package-Requires: ((emacs "26.1"))