+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.19] - 2022-11
 ### New
 - Add support for parsing the `BIT` and `VARBIT` datatypes.
+- Add support for parsing ARRAY datatypes.
 
 
 ## [0.18] - 2022-10-16