+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add support for parsing the `BIT` and `VARBIT` datatypes.
 - Add support for parsing ARRAY datatypes.
 - Add support for parsing RANGE datatypes (integer and numerical).
+- Add support for parsing HSTORE datatypes.
 - Add function `pg-cancel` to request cancellation of the command currently being processed
   by the backend.